Added arg to __init__ of BotStreamListener
This is experimental, and will probably not work, but the goal is to tac on the "bot" arg and line of code to __init__ of BotStreamListener, so that methods will have access to the enstantiated instance of the Bot class and it's methods.
so something like filehandle = self.bot.extractTL(tweet) will work from within the streamListener class, making things much easier and more flexible to code changes.