Fixed markov lib
Before, the function to return a chain was in __init__. This didn't make sense, as that gets called when the class is created and return values from that don't make it back to the caller, or... At least... Not the way I want them to.
So now __init__ accepts default values for NGram size and response length. If values for those options aren't given to getChain (the function that handles the hole chain making and random selection process), those defaults are used instead.